expressions: s/__in_fake_var_assign/is_fake_var_assign(expr)/
The problem with the __in_fake_var_assign variable is that it wasn't
taking inline functions into consideration.  So it would say that all
assignments in an inline function were fake when that's obviously wrong.

It meant that untracked variables were not handled correctly when the
function was inline.  But the DB looked fine so it was confusing.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>