user_data: just say socket information is not user data
This commit is pretty messed up.  Sorry.  I had wished the changes to
points_to_user_data were in a separate commit but it became complicated.
Also a lot of the changes to fixup_kernel.sh are probably not required now.
:(

How Smatch tracks user data across function boundaries has never worked
100%.  The problem is that we set "foo->bar" and record that in the
return_states table and whether it is user data depends on what we passed
into the function.  Ideally the user data code would say it depends on
if $->foo->bar was user data, but what it does instead is just give you
a flag to say "did you pass any user data to this function or not?".

So then if we pass one tiny user controlled variable it marks all the
returned data as user controlled.  And often it's like everything in
the skb is marked as user controlled.  It's hard to filter that stuff out
of the database because of recursion.

So instead of that, let's just say that the users cannot create a user
controlled socket data.  This is basically true, and it really helps cut
down on the number of bogus false positives.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>