iterator_used_outside_loop: check if bad list_for_each() iterator is used
The list_for_each() macros loop until they get back to the list head.

If the list iterator points at the list head then it can't generally be
used.  I have fixed a bunch of these bugs and there is a Coccinelle script
as well so all the remaining uses in the kernel are correct.  Still the
check is pretty useful for new code.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>