flow: fix parsing of min()/max() macros
The problem is that strip_expr() strips out the __builtin_choose_expr()
call and turns this function into a no-op.  However Smatch extra uses
strip_parens() so it doesn't parse this.  And function_hooks uses
strip_expr() so it doesn't parse this.

The result is the "foo = min();" does not get parsed and Smatch prints an
uninitialized variable warning.

Reported-by: Hans Verkuil <hverkuil@xs4all.nl>
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>