signed: silence false positives when comparing int vs ULONG_MAX
What often happens is that you have a condition like:

	if (uint_value > ULONG_MAX / 32)
		return -EINVAL;
	frob(uint_value * 32);

The condition is checking for integer overflows on 32 bit systems.  The
condition and the integer overflow are impossible on 64 bit systems.  This
is valid code and warnings about it are false positives so we do not want
to print a warning.

The key innovation in this patch is that it looks for ULONG_MAX on the
far left of the "ULONG_MAX / 32" expression.  We always start with
ULONG_MAX and then divide or subtract to get the upper bounds before an
integer overflow will happen.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>