helper: make macro_to_ul() slightly less bad
The macro_to_ul() is designed to turn a "#define FOO 2" and get a 2.  But
"#define __GFP_ZERO      ((__force gfp_t)___GFP_ZERO)" has the __force
annotation.  So skip over everything until you find the first number.

This is obviously terrible, but this function only is used for specific
macros so it works.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>