ssa: change how updates to ssa states happen
One problem with the SSA code is that say you have:

	p = alloc();
	...
	free(p);

Imagine the alloc() function has two success paths that means there are
two SSA states associated with "p".  So when we hit the free(p) function
we want to set both SSA states to &freed.  That didn't happen in the
original code.

What we can do instead is to return the ssa sm but change the sm->state
pointer to point to the state from the caller script.  Then use the
update_ssa_sm() function to update all the relevant SSA states.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>