unwind: mark all matches as undefined
I'm not really sure what this change does... But I'm trying to make my
diff with the released code smaller and this seems harmless and makes the
code smaller.

The mark_matches_as_undefined() function is supposed to silence false
positives.  What it does is if a resource is freed and we can't tie it to
a resource which we have marked as allocated then try to see if there is
something with a similar name and mark that as undefined.  It's supposed
to silence false positives.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>