ssa: fix a bug caused by recent cleanup
I don't actually know that this is caused by the recent cleanup, but it
definitely didn't help.  In olden code when there was just one ssa state
that was handled by a special case.  But then it was merged into the loop
and unfortunately the "tmp == sm" means that this would return NULL
instead of the correct state.

(The behavior changed as well so the cleanup was necessary anyway.  It's
not like I go deliberately introducing bugs.  Also cleanups are good
generally).

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>