db: improve is_kernel_error_path()
I wrote this code some time back and never merged it.  What I need to do
is make this a per project callback which is set in check_kernel.c.
However, I'm going to commit the code that I have for now and commit the
better code when I write it.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>