db: always call the returned info
Say a function does this:

struct my_struct *frob(struct whatever *p)
{
	p->foo->bar = user data;
	return p->foo;
}

We should record in the database that the parameter was set to user data
and we should also record that the function returns user data.  We can't
just return one or the other, because we don't know what the caller will
care about.

Ideally there would be a better, less duplicative way to handle this where
we could record just one entry that $1->foo and $(-1) are identical but for
now this is the only option.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>