state_assigned: new hook for states copied from one variable to another
One thing that I seem to implement over and over is tracking when a
state is copied from one variable to another.  What this code does is it
hooks into assignments, then if we have a state for the right hand side it
calls the hook.

This is sort of like smatch_ssa in a way and maybe using ssa is the way to
go so maybe this won't be used.  But I've been meaning to implement this
for a while.  I'm trying to commit code more often instead of writing it
and then forgetting why I implemented it.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>