db: fix returning container of code
The bug here is that the return_str is supposed to be record that the
return is a container_of() like "4096-ptr_max[(8<~$)]".  But if a function
was a direct container_of() return:

struct foo *to_foo(struct bar *p)
{
	return container_of(p, struct foo, bar);
}

The the container_of relationship is stored in the fake assign.  So we
need to get that earlier in the get_return_ranges_str().  Then the other
problem was that the strip_expr() at the top of call_return_state_hooks()
breaks get_fake_variable().

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>