function_hooks: fix handling of fake assignments for param/key code
The problem is this, imagine you have a param key of $(-1).  The
param_key_function() hooks into the function_call hooks, and not the
function assignment hooks.

One solution might be to check if the parameter was -1 and hook it into
the assignment hooks.  But that's not the approach that Smatch takes.
Instead it looks up the assignment expression by following the parent
pointers.  The problem with that is that it does not look up fake
parent assignments.

So "ret = frob(); return ret;" was handled but "return frob();" was not
handled.

Look up the fake parent assignment.  If this continues to be a problem then
maybe create a different hook for -1 parameters.

Reported-by: Elena Reshetova <elena.reshetova@intel.com>
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>