ssa: delete ssa_set_all_states()
This code was used to do a kind of half ssa thing.

	foo = bar;
	free(foo);

Then you could use this code to set both foo and bar to freed.  But it was
a hack because it said if we had an SSA that originally referred to bar
then "bar" is now freed.  It would need to look at the current SSA for
bar.

I thought about trying to fix this but I think that actually if you want
SSA you should just go all the way instead of half way.  Half way is not
a supported thing now.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>