unwind: cleanup and fix ordering
I was planning to create a new infrastructure so that I could do the
unwind RELEASE inserts before the PARAM_SET inserts.  But I decided to
first clean it up to use the add_return_info_callback() callback because
if I'm going create a new infrastructure then I want to make sure it's for
the newest stuff.

But woohoo!  This code already runs before PARAM_SET.  At some point I will
probably have to create add_return_info_callback_late() but not today.
Today, it's the weekend!

Anyway, the bug is that if you have:

	free(foo->bar);
	foo->bar = NULL;

Then we want to record that "foo->bar" was released before it was set.
Before it was recorded in the opposite order so it triggered a leak
false positive warning in the caller.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>