db: print container_of() string for known pointers
This is a thing with the mtag code where we give a known value to certain
pointers.  If we always pass a known mtag value and we return the
container of it, then the return string was the mtag for the container.
But we actually still want to record that it's the container as well.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>