check_unwind: fix ordering properly
I previously thought I had fixed the ordering because I changed how the
date was recorded in the DB.  Also my testing said it was fixed.  But
further testing showed it wasn't fixed.

When it comes to ordering, the type is what really determines the order,
not how it looks like in smatch_data/db/smdb.py output.  I should probably
update smdb.py to order it in the correct way as well...

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>