db: try split returns by success/fail earlier
There are a bunch of different ways that Smatch tries to split the returns
in meaningful ways.  I think the issue here was that I wanted to not
have so many returns where it gets split into negatives, zero and positives.
In the kernel, it should just be zero vs non-zero mostly.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>