unwind: add function param/key to ignore pointers
This was from tracking alloc_page().  There are some times where the code
does this:

	page = alloc_page(GFP_KERNEL);
	__skb_frag_set_page(frag, page);

After that point, the page is freed with the skb and it's not a leak if
we don't call free_pages() on error.  But __skb_frag_set_page() doesn't
release the page either so it's not a use after free if we use page later.

Now we can just mark it as an ignore situation and silence the warning.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>