struct_assignment: fix a bug generating member expressions
The struct assignment code is supposed to handle a couple things:

	memcpy(foo, bar, sizeof(*foo));
	memcpy(&foo, &bar, sizeof(foo));

And the same thing for assignments:

	foo = bar;

Where foo could either be a pointer or a struct.  Then we create a bunch
of fake assignments like "foo->a = bar->a;" or "foo.a == bar.a".  But
sometimes Smatch does crazy things and takes an expression like
foo = &bar and generates expressions like "foo->a = &bar->a" where
"foo->a = bar.a" was intended.

I noticed this when I was looking at fake assignments passing parameters.
These fake assignments are used for tracking user data so Smatch would say
"&bar->a" doesn't hold user data and the information was lost.

So make everything consistent and when we have pointer assignments
"p = foo" then convert that to "*p = *foo" for the fake assignments.  This
is what memcpy() already does.  So this removes a complicated special
case which was buggy before.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>