extra: handle the PARAM_DEREFERENCED hook earlier
One thing which happens is that we dereference a parameter and then set it.
The dereference should be processed first.  Otherwise if it's processed
second maybe we set the parameter to NULL, and then processing the
dereference means it can't be NULL so now it's an impossible state.

Also it makes no sense to process dereferences afterwards.  There is
nothing to be gained from that.  The PARAM_SET will hopefully say that it
must be valid and that's the end of that.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>