extra: clean up in values_fit_type()
I cannot see that this change should have any effect on runtime.  I think
it's just a clean up.  In this code we are worried about if negative values
from right side of an assignment will fit into the variable on the left
side.  The new code seems more readable to me.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>