points_to_user_data: set original pointer as well
This is part of a different bugfix, but I am pulling it into it's own
thing.  The truth is that even this small section is probably not the
right thing, now that I think about it.  But imagine you have code like:

frob(char *buf, char *end)
{
	char *p = buf;

	do {
		*p = user data;
	} while(p++ < end);
}

We want to say that "buf" points to user data at the end.  The problem is
that "p" does not point to "buf" at the end so this relationship is lost.

This change cannot hurt anything and can only make things better so I am
committing it.  However, I am not necessarily committing the rest of the
bug fix so my test case is broken for now.  And also I could easily write
a test case which used "buf++" instead of "p++" and that would be broken
too.  It requires some more thought.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>