spectre_second_half: clean up the code a bit
The problem was I was investigating a use after free in this code on my
devel tree.  And it seemed like "Oh.  We parse the inline, then free some
stuff on the stree, then there is a use after free."  So I made the change
and the problem was fixed.

But as I write the code, I realize "Hey, we shouldn't free anything on
the stree until we are done parsing all the inlines.  Also it's fine to
have states from inlines on the stree because we're not going to look them
up by mistake.  So the code should work as-is."

However conceptually, this patch is a clean up.  Let's keep not mix states
from one function to the other.  It also does fix the bug which is good.
If the real bug is still there probably it will trigger a different crash
and I will fix it properly next time.

[ edit: Yup.  The use after free was caused by debug code ]

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>