expressions: fix a timeout test
This code expected "count" to be zero after a timeout, but it's actually
set to -1.

I don't remember writing this fix so I can't say the impact.  :(  I'm
trying to do better and commit code as soon as I can now even if I have to
fix or revert it later.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>