expressions: introduce new helpers for allocating permanent expressions
I'm moving more and more to creating fake assignments.  The problems is
that when these fake expressions are freed at the end of parsing a
function.  So then next time when we parse the function as an inline, there
are freed fake expressions pointers hanging around.

Smatch kind of worked around this by checking cur_func_sym->parsed in
get_state_expr() and ignoring fake expressions.  But that is an ugly hack.
There is no reason why we can't just re-use the old fake expressions for
the second parse.

This patch just introduces some helper functions to create long lived
allocations.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>