preempt: fix how preempt_count() is handled
The impetus for this change is a devel tree which changes how functions
are handled.  The kernel code is:

	WARN_ON(in_interrupt());

I wasn't anticipating that's how in_interrupt() would be used.  Normally
it's used like:

	if (in_interrupt())
		// use ATOMIC code

You would think that in_interrupt() would say the true side to atomic and
the false side to not and they would merge into not atomic.  But what
happens is that in_interrupt() returns in the DB do not break down
cleanly into true false but are instead just unknown.  So both call backs
are called and it sets it first to not atomic and then to atomic so the
result is that it's in an atomic state.

Use return_implies_exact() to avoid this.

The second observation is that this code is mostly for silencing scheduling
in atomic bugs.  We want to err on the side of marking things as not
atomic.

Calling preempt_count() cannot make the code atomic if it wasn't atomic to
begin with so update match_preempt_count_non_zero() to match.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>