flow: create fake assignments for function calls
Imagine you have code like:

	foo = 1 + func(bar);

The problem is that a Smatch modules generally want to store states with
a variable.  So where do we store that return values from func()?  Some
modules like smatch_extra.c and smatch_user_data.c do some custom weirdness
where they create a "return %p" variable and store it there.  But that is
non-standard.  It also doesn't work well with the param/key API which
really want a variable to store -1 parameter values.

The solution is to create a fake variable and a fake assignment so it
becomes:

	__fake_123 = func(bar);
	foo = 1 + __fake_123;

That is what this change does.

There are some TODOs remaining.  The big one is that comparisons are not
handled because then the validation/sm_compare15.c test breaks.  I plan to
add support for this.  The other thing is that smatch_extra.c and
smatch_kernel_user_data.c still refer to "return %p" and I need to delete
that code.

But I am trying to be better about committing code as soon as I can so I'm
committing this code now.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>