function_hooks,check_unwind: fix how return_implies functions are handled
This patch only affects what happens when you don't have the cross function
DB so I hadn't seen the bug until it was reported.

Back in the day, I used say that, for example, spin_trylock() return 1
then it implied that the state was locked and if it returned 0 then the
state was unlocked.  But had some subtle side effects and really I was only
interested in when the state was changed to locked.  So I only needed to
implement the first part, that return 1 implies locked.

So then what happens here is that we iterate through the hooks and say
okay it returns 1.  But the return 0 was not recorded.

We need to say at the end, "which possible values are not handled" and then
create a nothing happens return for "no change" on this path.  An additional
thing we could do would be to create a way to specify all the possible
return for a function.

The way that this bug was exposed was:

drivers/media/platform/ti/davinci/dm644x_ccdc.c:833 dm644x_ccdc_probe()
warn: 'ccdc_cfg.base_addr' from ioremap() not released on lines: 833.

The problem was again that Smatch thought that ioremap() cannot fail.
But then an additional problem is that although the check_unwind.c check
has checks for the impossible path, it doesn't check at the goto.  By the
time it gets the return and checks then the path has merged with a possible
path.  This is something that could be checked in a pre-merge hook.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>