refcount_info: replacement for check_atomic_inc_dec.c
This just provides easy to use call backs so other checks can track when
a reference is incremented or decremented.  I have written a some but
they're not ready to be published yet.  But I want to have this released
so other people (real people from work, not my imaginary friends) can use
it and look at it.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>