estate: "capped state + small state" leads to a capped state
The question here is say we have one side of a merge where a variable is
capped to a small literal and the other side is capped to an unknown value.
Intuitively the result should be saved as capped to an unknown.  I have
chosen 99 as the max literal.

This is a bit hokey, but I feel like it's better than the previous
behavior.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>