uninitialized: fix checking for fake calls and don't save global states
The real reason for this change is that I don't want to store global
states.

	if (!cur_func_sym)
		return;

But unfortunately, that change was tangled in with some other changes
related to checking for fake calls.  So it all got committed together.

This was part of the change which makes the "uninitialized special assign"
warning work.  Before I had thought we just don't have any bugs related
to special assigns...  But that code is not ready for prime time so I'm
not going to fix it 100%.  It's almost there though...

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>