logical_instead_of_bitwise: add some more warnings
The important difference is the more flexible check_is_mask().

The problem is trying to figure out which code is bitwise and one trick
that I tend to use a lot is assuming that 0-1 is boolean.  So, for example,
if (foo > bar) is boolean or maybe a function returns 0-1 so that's
probably boolean.

One recent bug used "foo && BIT(0)".  The BIT(0) value is 1 but because of
the left shift we know it's bitwise and not boolean.  I have instead
added a check on the left hand side because sometimes that can be a
comparison.

Then this patch adds some other warnings as well.  But only the "should
this be a bitwise negate mask?" ever triggers and I decided to disable that
warning unless --spammy is enabled.  It doesn't trigger many warnings but
I can't understand most of that code so I'm not sure if it's correct or
not.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>