kernel_user_data: set points to user data for dereferenced pointers
Imagine you have:

	*p = get_user_data();

Then that means "p" points to user data.  Normally setting points to
user data is not required but where it is is when we modify "p" later.

	*p++ = skb->data[0];

I had considered a more narrow way of only setting points to user data
when there is a ++ involved but I never wrote that code so I'm going to
commit this.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>