conditions: set parent expression for conditional conditions
If you have "foo = bar ? 1 : 2;" then the parent statement for "bar" is
not set.  The parent should always be set.  This is a bug.

Also if you have a statement like "foo = bar?: 2;" then we create a fake
assignment and set "foo = bar;" and I think the parent expression for that
can be "bar?:2"...  To be honest, I'm not sure about this but it feels like
sort of the right thing.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>