flow: fix handling of faked function assignments
The bug here is that if we parse a function twice (because it is parsed
inline) then the second time through we need to handle the fake assignments
for function calls as well.

I wanted the code to work like this originally.  But there was a bug where
it parsed things recursively in a forever loop until it ran out of memory
and crashed.  So then I guess I just added a return, and didn't think about
it again.

What ends up happening is that now we have parse those calls.

The fix is if we have already created a fake assignment then re-use that
but add a test for recursion.  It's pretty easy to test for recursion
because gen_fake_function_assign() will be called in a tight loop.  It's
sufficient to just store the current fake assignment that we are parsing.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>