param_cleared: rename and re-order PARAM_CLEAR
I was going to leave PARAM_CLEAR and transition in a different way, but I
decided the changes were incompatible and it wasn't worth it.

The PARAM_CLEAR type was in the wrong place.  It needs to be after
PARAM_LIMIT but before PARAM_SET.  I just renamed it to BUF_CLEAR because
I started with that as a transitional name, and then decided I liked it.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>