missing_break: delete this check
Originally, Smatch tried to determine if there was a break missing by
looking for unused assignments.  Dead stores or whatever.  But these days
we look at it like a bug if there is any fall through without a proper
fall through annotation.  That's a better rule and simpler to implement.

This Smatch check uses a lot of resources.  Other static checkers and GCC
have checking for missing fall through annotations so this isn't needed
any more.  Delete it.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>