param_cleared: do not carry around tons of void pointer information
Smatch tracks parameter values by strings "$->foo".  But if you have a
void pointer, then maybe "$->foo" doesn't make sense if "$" is void.
But I always just thought it was harmless to carry that data around and
maybe it would mean something to the right caller.

The thing is that now the Smatch database has become too large so something
has to go.  So I'm not going to carry this stuff any more.  Just mark it
as modifying "*$" and be done.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>