param_clear: smush huge unions into a single param clear
Smatch tracks variables by their printed names instead of by their memory
locations.  So unions are a problem.  Sometimes you get a function which
sets hundreds of union members and sub members to unknown and that's a lot
of data.  It's kind of useless and misleading data as weall.

Just combine it all into one "*$" BUF_CLEARED and be done.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>