param_used: do not store PARAM_USED states when the variable is PARAM_CLEAR
This is like param_was_set_var_sym() but for if the whole buffer was
cleared.  Saying the "whole buffer" is in not necessarily entirely
accurate because the BUF_CLEAR can be set when the whole buffer is not
100% set.  It's an approximation.

But the point is we don't want to carry these PARAM_SETs and the PARAM_USED
data is even less useful.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>