param_limit: don't store some PARAM_LIMIT information for dereferenced pointers
The issue is, param_limit doesn't actually look at if statements.  It looks
at changes to SMATCH_EXTRA states that are not based on setting variables.

So when we dereference a pointer smatch extra says "Okay.  That means it
can't be NULL."  Which is fine.  But that information gets passed back
and back to all the callers.

So then maybe a then maybe the function is only called sometimes so we have
a things where we have to merge unknown + valid.  Which ends up being a lot
of states.

The right fix is to re-write param_limit to ignore dereferences.  But that's
quite complicated.  Another fix is to say that if the variable name is
just really long like "$->i_sb->s_bdev->bd_inode->i_mapping->foo->bar"
then we can safely ignore the implications from it.

This does help speed up parsing some files quite a bit.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>