param_limit: try ignore dereferences harder
Quite a few of PARAM_LIMIT entries are generated because of dereferences.
And it's true, that if you return on this path it means that $->foo->bar
must be non-NULL.  But at the same, time storing that information uses
quite a bit of space. And whenever we come across a PARAM_LIMIT then
smatch tries to figure out the implications of it.  And the information is
not really useful because it doesn't say if the pointer is checked before
dereferencing or not.

So I initially thought, let me add a nomod hook which stores every extra
state modification which comes from an if statement or a PARAM_LIMIT and
we can ignore all the PARAM_LIMIT which don't come from an if statement.

But I accidentally typed it wrong and made it a mod hook instead of nomod
hook.  In trying to debug it, I wrote the code to be more and more careful.
For example, I thought, maybe there are places which call
set_state(SMATCH_EXTRA, ...) directly.  So then I said, let me add two
lists, one with a list of things we can ignore and one with a list of
things we have to pay attention to.

And in the end, the extra checking is required because of all the different
names a variable and have.  I kind of like the extra checking as well.
It's cautious.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>