param_clear: record fewer void pointer states
The caller side already tries to ignore these, but we shouldn't even put
them into the DB.  It's the same reasoning as forunions, except even more
so.

When I was testing this against mcp251xfd_regmap_nocrc_update_bits() I
thought that it wasn't working because we were still recording 133 states
on the error path.  But then I looked original DB and there were recording
1131 states.  So it does work-ish.

But it's still a bit crazy.  The rule for BUF_CLEAR is that we don't store
states where the end result is just unknown.  But in this case even for
those 133 states, there isn't really a huge value in recording them.

Anyway, baby steps.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>