fix infinite loop when expanding __builtin_object_size() with self-init vars
expand_object_size(), used to expand __builtin_object_size(),
recursively try to get the parent initializer. This fails miserably
by looping endlessly when the object is a self-initialized variable.

For the moment, fix this in the most obvious way: stop the recursion
and do not expand such variables.

Note: I wouldn't be surprised if these self-initialized variables create
      other problems elsewhere. Maybe we should remove their initializer
      and somehow mark them as "do not warn about -Wuninitialized"
      (well, there is no such warnings *yet*).

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>