user_data/host_data: avoid DB lookups handling fake assignments
This is a part of my on going efforts to speed up Smatch builds.  If you
have code like:

	foo = bar;

Then Smatch translates that into:

	foo->a = bar->a;

Then the user_data and host_data try to calculate the user or host values
of bar->a.  To do that we have to calculate the value of bar->a and swap
int the user value for any extra values.  To calculate the extra values we
have to do a DB look up.  Then normally we say, "No user data here, discard
the results."

In this code, there is no need really to do a DB lookup.  But a simple
assignment can result in 50 DB round trips.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>