extra: handle more parameter dereferences
Imagine you have:

	lock(&foo->lock);

And the lock() function dereferences the parameter, then we know that "foo"
can't be NULL.  Previously I guess we would say that &foo->lock is not NULL
but that's not very useful.  I've been running with this code for a long
time so I guess it must work.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>