param_limit: ignore dereferences.  Harder and Hackier
I had tried to ignore dereferences in terms of PARAM_LIMIT before.  But
there were two problems.  1)  The "expr->type == EXPR_DEREF" check wasn't
correct.  It should be looking up the parent and testing for
"parent->type == EXPR_PREOP && parent->op == '*'".  2)  Even when you
don't record the PARAM_LIMIT in the DB, you still record a
PARAM_DEREFERENCE and that doesn't have an expr associated with it so it
still becomes a PARAM_LIMIT in the caller.

The fix it to just hard code into smatch_extra.c to ignore dereferences.
It's suitably ugly for a hack that I hope does not get used anywhere else.

This patch is mostly motivated by parsing vfs_fallocate().  The
vfs_allocate() function dereferences "file" on the first line and then it
has 18 returns on after the other and each return was a PARAM_LIMIT "file"
is non-NULL.

It takes a long time to parse all these because the caller tries to figure
out the implications of each PARAM_LIMIT.  Before this patch parsing
fs/ksmbd/vfs.c used to take 132 seconds but now it takes 49 seconds.

#NoLIMITS #HardCORE

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>