free_strict: use param/key API
This change is not supposed to change run time at all.  The param/key API
is more powerful and flexible and it might be useful to talk about freeing
struct members.

The truth is that I wanted to talk about freeing container_of()s but the
Smatch format for this is not useful for human beings.  I need to re-think
that format.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>