select_type: new check for propagating negatives to u32 to s64
This is for bugs like:

	s64_val = condition ? u32_val : -EINVAL;

The bug is that s64_val will never be set to -EINVAL.  The u32_val will
type promote it to (u32)-EINVAL instead which is a high positive value.

These bugs aren't super common in the kernel and we fixed most of them in
2021 when this check was written.  There are 5 remaining warnings: 2
real, 1 real but harmless and 2 false positives.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>